use double instead of float in events

+++ shared/agoclient.cpp
@@ -786,7 +786,7 @@ bool agocontrol::AgoConnection::emitEvent(const char *internalId, const char *ev
     content["uuid"] = internalIdToUuid(internalId);
     return sendMessage(eventType, content);
 }
-bool agocontrol::AgoConnection::emitEvent(const char *internalId, const char *eventType, float level, const char *unit) {
+bool agocontrol::AgoConnection::emitEvent(const char *internalId, const char *eventType, double level, const char *unit) {
     Variant::Map content;
     content["level"] = level;
     content["unit"] = unit;
